refactor: moved ColorModeHandler to /root directory
